Linux support

+++ Profile.ps1
@@ -11,7 +11,7 @@ Import-Module posh-git
 $EnvDirectory = $PSScriptRoot
 $ReposDirectory = Split-Path $EnvDirectory
 $EnvPSModuleDirectory = Join-Path $EnvDirectory 'PSModules'
-Import-Module -Force "$EnvPSModuleDirectory\Paths\Paths.psm1"
+Import-Module -Force (Join-Path $EnvPSModuleDirectory 'Paths' 'Paths.psm1')
 
 Register-PathElement -PathEnvironmentVariable 'PSModulePath' -PathElement @(
     $EnvPSModuleDirectory
@@ -24,17 +24,19 @@ foreach ($ModuleDirectory in (Get-ChildItem $EnvPSModuleDirectory)) {
 while (Test-Path alias:prompt) { Remove-Item alias:prompt }
 function global:prompt { Write-PromptEx }
 
-while (Test-Path alias:cd) { Remove-Item alias:cd }
-Set-Alias -Force -Name cd -Value Set-LocationEx
-
-New-Alias -Force -Name 'which' -Value 'Get-Command'
-
 New-Alias -Force -Name 'g' -Value 'git'
 New-Alias -Force -Name 'y' -Value 'yarn'
 
-New-Alias vs2017 'C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\Common7\IDE\devenv.exe'
-New-Alias vs2019 'C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\Common7\IDE\devenv.exe'
-New-Alias vs 'C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\Common7\IDE\devenv.exe'
+if ($global:IsWindows) {
+    while (Test-Path alias:cd) { Remove-Item alias:cd }
+    Set-Alias -Force -Name cd -Value Set-LocationEx
+
+    New-Alias -Force -Name 'which' -Value 'Get-Command'
+
+    New-Alias vs2017 'C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\Common7\IDE\devenv.exe'
+    New-Alias vs2019 'C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\Common7\IDE\devenv.exe'
+    New-Alias vs 'C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\Common7\IDE\devenv.exe'
+}
 
 function global:cdr { cd $ReposDirectory }
 function global:cdasc { cd (Join-Path $ReposDirectory 'axe-sarif-converter') }
@@ -43,3 +45,5 @@ function global:cdenv { cd (Join-Path $ReposDirectory 'dbjorge-env') }
 function global:cdjorbs { cd (Join-Path $ReposDirectory 'jorbs-spire-mod') }
 function global:cdweb { cd (Join-Path $ReposDirectory 'accessibility-insights-web') }
 function global:cdwin { cd (Join-Path $ReposDirectory 'accessibility-insights-windows') }
+
+cdr
